+++ src/main/java/pl/adamsiedlecki/ohm/dto/HumidityDto.java
@@ -2,7 +2,6 @@ package pl.adamsiedlecki.ohm.dto;
 
 import lombok.Value;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Value
@@ -12,5 +11,5 @@ public class HumidityDto {
     LocalDateTime time;
     long stationId;
     String stationName;
-    BigDecimal humidity;
+    double humidity;
 }
